MySQL

+++ controller/post.mjs
@@ -22,22 +22,29 @@ export async function getPostById(req, res, next) {
 
 //포스트를 생성하는 함수
 export async function createPost(req, res, next) {
-  const { userid, name, text } = req.body;
-  const posts = await postRepository.create(userid, name, text);
+  const { text } = req.body;
+  const posts = await postRepository.create(text, req.useridx);
   res.status(201).json(posts);
 }
 
 // 포스트를 업데이트 하는 함수
 export async function updatePost(req, res, next) {
-  const id = req.params.id;
-  const text = req.body;
-  const posts = await postRepository.update(id, text);
+  const idx = req.params.id;
+  const { text } = req.body;
+  const posts = await postRepository.update(idx, text);
   res.status(201).json(posts);
 }
 
 // 포스트 삭제 함수
 export async function deletePost(req, res, next) {
   const id = req.params.id;
-  const posts = await postRepository.remove(id);
-  res.status(201).json(posts);
+  const post = await postRepository.getById(id);
+  if (!post) {
+    return res.status(404).json({ message: `${id}의 포스트가 없습니다` });
+  }
+  if (post.useridx !== req.useridx) {
+    return res.sendStatus(403);
+  }
+  await postRepository.remove(id);
+  res.sendStatus(204);
 }
